Algorithm/DP/LC139 单词拆分_test.go
package DP

import (
	"testing"
)

func Test_wordBreak(t *testing.T) {
	type args struct {
		s        string
		wordDict []string
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		{"1", args{s: "leetcode", wordDict: []string{"leet", "code"}}, true},
		{"2", args{s: "applepenapple", wordDict: []string{"apple", "pen"}}, true},
		{"3", args{s: "catsandog", wordDict: []string{"cats", "dog", "sand", "and", "cat"}}, false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := wordBreak2(tt.args.s, tt.args.wordDict); got != tt.want {
				t.Errorf("wordBreak() = %v, want %v", got, tt.want)
			}
		})
	}
}
